move optimizer to _optimizer to fix pyre bug.
