+++ torch/optim/_multi_tensor/adadelta.py
@@ -1,5 +1,5 @@
 import torch
-from ..optimizer import Optimizer
+from .._optimizer import Optimizer
 
 class Adadelta(Optimizer):
     """Implements Adadelta algorithm.
