+++ torch/optim/_multi_tensor/adam.py
@@ -1,6 +1,6 @@
 import math
 import torch
-from ..optimizer import Optimizer
+from .._optimizer import Optimizer
 
 class Adam(Optimizer):
     r"""Implements Adam algorithm with multi tensor APIs.
