+++ torch/optim/_multi_tensor/asgd.pyi
@@ -1,5 +1,5 @@
 from typing import Tuple
-from ..optimizer import _params_t, Optimizer
+from .._optimizer import _params_t, Optimizer
 
 class ASGD(Optimizer):
     def __init__(self, params: _params_t, lr: float=..., lambd: float=..., alpha: float=..., t0: float=..., weight_decay: float=...) -> None: ...
