+++ torch/optim/adadelta.py
@@ -1,5 +1,5 @@
 import torch
-from .optimizer import Optimizer, _params_t
+from ._optimizer import Optimizer, _params_t
 
 
 class Adadelta(Optimizer):
