+++ torch/optim/adam.py
@@ -2,7 +2,7 @@ from typing import Tuple
 
 import torch
 from . import functional as F
-from .optimizer import Optimizer, _params_t
+from ._optimizer import Optimizer, _params_t
 
 
 class Adam(Optimizer):
