+++ torch/optim/adamax.py
@@ -1,7 +1,7 @@
 from typing import Tuple
 
 import torch
-from .optimizer import Optimizer, _params_t
+from ._optimizer import Optimizer, _params_t
 
 
 class Adamax(Optimizer):
