+++ torch/optim/asgd.py
@@ -1,6 +1,6 @@
 import math
 import torch
-from .optimizer import Optimizer, _params_t
+from ._optimizer import Optimizer, _params_t
 
 
 class ASGD(Optimizer):
